Use useSearchCourse custom hook
