+++ client/src/shared/utils/index.ts
@@ -2,3 +2,4 @@ export * from './axios';
 export * from './toastify';
 export * from './interface';
 export * from './date';
+export * from './debounce';
